requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.12
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
redis==6.0.0
rich==13.9.4
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
